backend/pyproject.toml
[project]
name = "ai-studio-backend"
version = "1.0.0"
description = "AI Studio Backend API - FastAPI application for interior design render generation"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Adityo Nugroho", email = "adityo.nugroho.id@gmail.com" }
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "google-genai>=1.47.0",
    "pillow>=11.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
